53 refactoring of database to front end systems (#57)
* DTO Update

Renamed ViewModels to DTOs.
Cut down DTOs to be smaller.
Added DTOs for DTO collections.

* Model Variable Nullability

Just making sure some ICollections are nullable, as it isn't a given that there is something there.
And nullability for the user avatar, as we could instead have a default avatar when an avatar is lacking.

* Construction and Renaming

Added constructors to all DTOs, and renamed DTOs containing collections to avoid a previous mistake, aka dyslexia.

* DTO and Controller update

DTO collection constructors and Controller updated accordingly.

* Improved DTO and Controllers, and removed Like

Improved the DTOs by adding the ID of the layer above them for quicker references and call queries. Updated controllers to the new refactored system. And removed the Like object as it was not needed. It is now baked into comment and post.

* Fixes and Loading Overhaul

Fixed deserialization error by adding empty constructors to all DTOs, removed Automapper, overhauled Repository pattern to explicitly load more children, updated models referencing each other for the relational mapping, added pagination on a couple controllers, and added the beginning of a generic repository project.

* Structure update

Fixed errors, and improved models and DTOs to work with getting and posting requests.

* Renamed variables

Renamed ViewModel to Data, and DataCollections.
Fun Fact: There was exactly 100 variables named ViewModel.