Like update (#73)
* Like Update

Like functionality added to controllers.

* Like prototype and User Update

Likes have been implemented in a limited scope. When calling a GET for s ingle post or comment, it will check if the current user have liked the post/comment or not.
Also, now user ID is tied to firebase user_id.
Also also, Neighborhood variable for categories is nullable.
User is officially the top of the hierarchy.

* Removed Commented-out code

As requested.