Pagination Update
Removed redundant Pagination functions as I found a way to include it in the basic GetById function.
Updated controllers to work with pagination.
Removed the parameters for page and size that was used for pagination, as those would only need to be set in the func parameter in the controllers when calling the function in itself.