105 direct repository pageination (#106)
* Async Update

Async rework to repository, service, and controllers.

Some null warning removed, and arrays replaced with ICollections.

* Removed CancellationToken

Was not found necessary as the project did not have massive calls that would require a possible cancelation.

* Pagination Update WIP

Getting pagination is more complicated than anticipated.
Here is an early try. WIP

* Removed Redundant null check

* Minor syntax fix

Removed redundant class mentions.

* Changed BadRequest

Changed BadRequest to Unautherized on relevant parts.

* Proof of concept Pagination

Found the way to satisfy the demand to insert a pagination function into the include function while keeping the deeper include function.
Also found a bug from the previous update to develop, and have fixed it.

* Pagination Update

Removed redundant Pagination functions as I found a way to include it in the basic GetById function.
Updated controllers to work with pagination.
Removed the parameters for page and size that was used for pagination, as those would only need to be set in the func parameter in the controllers when calling the function in itself.

* Removed Migration