Server / Client post and comment DTO (#113)
* Async Update

Async rework to repository, service, and controllers.

Some null warning removed, and arrays replaced with ICollections.

* Removed CancellationToken

Was not found necessary as the project did not have massive calls that would require a possible cancelation.

* Pagination Update WIP

Getting pagination is more complicated than anticipated.
Here is an early try. WIP

* Removed Redundant null check

* Minor syntax fix

Removed redundant class mentions.

* Changed BadRequest

Changed BadRequest to Unautherized on relevant parts.

* Proof of concept Pagination

Found the way to satisfy the demand to insert a pagination function into the include function while keeping the deeper include function.
Also found a bug from the previous update to develop, and have fixed it.

* Pagination Update

Removed redundant Pagination functions as I found a way to include it in the basic GetById function.
Updated controllers to work with pagination.
Removed the parameters for page and size that was used for pagination, as those would only need to be set in the func parameter in the controllers when calling the function in itself.

* Removed Migration

* Implement Client and Server post DTO

Added a new Client post DTO for transfer of data in a smaller package, for ease of development.
Adjusted the now named Server post DTO.
Implemented this new feature into the post controller.

* Server/Client Post and Comment DTOs

Implemented Server/Client Comment DTOs, as they had a familiar usecase to the Server/Client User DTO.

* Naming Quickfix

* UserController Hotfix

Fixed a duplicate tracking issue in the repository from multiple of the same neighborhood were being tracked.

* Reduced Client/Server Post and Comment DTO

Removed ID, as that is already a part of the HTTP request.