Assign board members and related changes (#125)
* Added calls to assign and unassign board members

Also added messages to errors.

* Fixed deleting neighborhood with users and fixes

Can now delete a neighborhood with users, their neighborhood id's will be null and their roles will be removed.
Added authorization for only board members to edit/delete neighborhoods, as well as needing to be a member of the neighborhood.
Added restriction to not be able to make a new neghborhood while being a member of one.
Added cascading delete to a few relationships to make sure things clean up as they should.
Also added messages to errors.

* Fixed get users from neighborhood not having roles

* Added author/admin check to edit/delete posts

Also added messages to errors.

* Added author/admin check to edit/delete comments

Also added messages to errors.

* Added admin check to creating/editing/deleting categories

Also added messages to errors.